# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WatchIT Tower Monitoramento — API
  version: "1.0.0"
  description: |
    API REST do WatchIT Tower Monitoramento — dashboard local de monitoramento de infraestrutura de TI.

    ## Autenticação

    A maioria dos endpoints exige uma sessão NextAuth ativa (cookie de sessão JWT).
    Faça login em `POST /api/auth/callback/credentials` e inclua o cookie `next-auth.session-token`
    nas requisições subsequentes.

    ### Exceções (sem autenticação de sessão)
    - `GET /api/health` — público, sem autenticação
    - `POST /api/links/{id}/up` e `POST /api/links/{id}/down` — autenticados por HMAC-SHA256 via header `x-webhook-token` ou query string `?token=`

    ## Rate Limiting
    O endpoint de login (`POST /api/auth/callback/credentials`) tem rate limit de
    **10 tentativas por IP em 15 minutos**.
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Cookie de sessão JWT gerado pelo NextAuth após login bem-sucedido.

    webhookToken:
      type: apiKey
      in: header
      name: x-webhook-token
      description: |
        Token HMAC-SHA256 gerado com `WEBHOOK_SECRET + linkId`.
        Pode ser enviado como header `x-webhook-token` ou query string `?token=`.
  schemas:
...
    DeviceType:
      type: string
      enum: [MIKROTIK, DVR, CAMERA, OTHER, UNIFI_AP]
      description: Tipo do dispositivo.
...
    BulkDeviceInput:
      type: object
      description: |
        Criação em massa de dispositivos por faixa de IP.
        O nome de cada device será `"{name} {último octeto}"` (ex: "Camera 101" para o IP .101).
        Máximo de 254 dispositivos por operação.
      required: [ipStart, ipEnd, name, type]
      properties:
        ipStart:
          type: string
          format: ipv4
          example: "192.168.2.100"
          description: Primeiro IP da faixa (inclusive).
        ipEnd:
          type: string
          format: ipv4
          example: "192.168.2.150"
          description: Último IP da faixa (inclusive).
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Camera"
        type:
          $ref: "#/components/schemas/DeviceType"
        location:
          type: string
          maxLength: 200
        pingEnabled:
          type: boolean
          default: true
        httpEnabled:
          type: boolean
          default: false
        httpPort:
          type: integer
          minimum: 1
          maximum: 65535
          nullable: true
        checkInterval:
          type: integer
          minimum: 10
          maximum: 3600
          default: 60
...
security:
  - sessionCookie: []

# ─── Paths ──────────────────────────────────────────────────────────────────
paths:
...
  /api/devices/bulk:
    post:
      tags: [Dispositivos]
      summary: Criar dispositivos em massa (por faixa de IP)
      description: |
        Cria múltiplos dispositivos de uma vez especificando uma faixa de IP (`ipStart` a `ipEnd`).
        O nome de cada dispositivo será `"{name} {último octeto do IP}"` — ex: nome `"Camera"` + IP `192.168.2.101` = `"Camera 101"`.

        IPs já existentes são ignorados (`skipDuplicates: true`).
        Limite de 254 dispositivos por operação.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeviceInput"
            example:
              ipStart: "192.168.2.100"
              ipEnd: "192.168.2.120"
              name: "Camera"
              type: "CAMERA"
              location: "Estacionamento"
              pingEnabled: true
              checkInterval: 60
      responses:
        "201":
          description: Criação concluída. Retorna contagem e lista de IPs criados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: integer
                    description: Número de dispositivos efetivamente criados (IPs duplicados não contam).
                    example: 21
                  ips:
                    type: array
                    items:
                      type: string
                    description: Lista de IPs criados.
                    example: ["192.168.2.100", "192.168.2.101"]
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
...